requirements.txt
faiss-cpu
langchain-community
langchain-core
langchain-ollama
pypdf
python-dotenv
tqdm
